test: extract a shared ffmpeg helper for the fixture generators
fixtures.mjs and library-gen.mjs each carried their own copy of the
bundled-ffmpeg path, the spawn wrapper, and the sine-tone encode args.
Pull them into test/helpers/ffmpeg.mjs (BUNDLED_FFMPEG,
assertFfmpegAvailable, runFfmpeg, encodeTone) so the two can't drift.

Side benefit: library-gen.mjs now inherits the race-safe temp-file +
atomic-rename encode that fixtures.mjs gained earlier — it previously
wrote ffmpeg's output straight to the final path. The bundled-binary
path also now lives in one place, ready for an env override when CI
needs a system ffmpeg.

No public-API or behavior change for the suite: identical ffmpeg args,
same exports (ensureFixtures / FIXTURE_SUMMARY / mkSpec / generateLibrary
/ DEFAULT_FFMPEG_PATH).
