+++ test/helpers/fixtures.mjs
@@ -9,10 +9,9 @@
  */
 
 import fs from 'node:fs/promises';
-import { existsSync } from 'node:fs';
 import path from 'node:path';
-import { spawn } from 'node:child_process';
 import { fileURLToPath } from 'node:url';
+import { assertFfmpegAvailable, encodeTone } from './ffmpeg.mjs';
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
 const REPO_ROOT = path.resolve(__dirname, '..', '..');
@@ -33,49 +32,10 @@ const FIXTURES = [
   { artist: 'Vosto',  album: 'Untitled EP',       year: null, genre: null,         disc: 1, track: 1, title: 'Sketch 1' },
 ];
 
-const BUNDLED_FFMPEG =
-  process.platform === 'win32' ? path.join(REPO_ROOT, 'bin', 'ffmpeg', 'ffmpeg.exe')
-                               : path.join(REPO_ROOT, 'bin', 'ffmpeg', 'ffmpeg');
-
-// Fixtures are encoded with mStream's bundled ffmpeg, which is gitignored.
-// A fresh git worktree won't have it, so fail with an actionable message the
-// first time we actually need to encode (rather than a cryptic spawn error
-// deep into a run). Checked once — if every fixture is already cached the
-// suite runs fine without ffmpeg present.
-let ffmpegChecked = false;
-function assertFfmpegAvailable() {
-  if (ffmpegChecked) { return; }
-  if (!existsSync(BUNDLED_FFMPEG)) {
-    throw new Error(
-      `ffmpeg not found at ${BUNDLED_FFMPEG}. Test fixtures are generated with ` +
-      `mStream's bundled ffmpeg (gitignored). In a fresh git worktree, copy ` +
-      `bin/ffmpeg/ from your main checkout before running the suite — see test/README.md.`,
-    );
-  }
-  ffmpegChecked = true;
-}
-
-// Monotonic counter for temp-file names so two encodes in one process never
-// collide on the same temp path (see encode() for why we rename).
-let tmpSeq = 0;
-
 function relPathFor(f) {
   const trackNum = String(f.track).padStart(2, '0');
   const safe = s => s.replace(/[/\\:*?"<>|]/g, '_');
   return path.join(safe(f.artist), safe(f.album), `${trackNum} - ${safe(f.title)}.mp3`);
-}
-
-function runFfmpeg(args) {
-  return new Promise((resolve, reject) => {
-    const p = spawn(BUNDLED_FFMPEG, args, { stdio: ['ignore', 'ignore', 'pipe'] });
-    let stderr = '';
-    p.stderr.on('data', d => { stderr += d.toString(); });
-    p.on('error', reject);
-    p.on('exit', code => {
-      if (code === 0) { resolve(); }
-      else { reject(new Error(`ffmpeg exited ${code}: ${stderr.slice(-500)}`)); }
-    });
-  });
 }
 
 async function encode(outPath, f, fixtureIndex) {
@@ -105,31 +65,7 @@ async function encode(outPath, f, fixtureIndex) {
   // content correctly share one audio_hash, but we want distinct content
   // here so per-track state stays per-track.
   const freq = 220 + fixtureIndex * 40;  // 220, 260, 300, … Hz
-  // Encode to a unique temp file, then atomically rename into place. Test
-  // files run in concurrent child processes, so on a cold checkout several
-  // may try to materialise the same fixture at once; writing ffmpeg's output
-  // straight to outPath would let a scanner read a half-written MP3. Rename is
-  // atomic within the directory, so readers only ever see a complete file. The
-  // `.mp3` suffix is preserved so ffmpeg still infers the mp3 muxer.
-  const tmp = path.join(
-    path.dirname(outPath),
-    `.tmp-${process.pid}-${tmpSeq++}-${path.basename(outPath)}`,
-  );
-  try {
-    await runFfmpeg([
-      '-nostdin', '-y', '-loglevel', 'error',
-      '-f', 'lavfi', '-i', `sine=frequency=${freq}:sample_rate=44100:duration=1`,
-      '-ac', '2',
-      '-c:a', 'libmp3lame', '-b:a', '64k',
-      ...metaArgs,
-      '-id3v2_version', '3',
-      tmp,
-    ]);
-    await fs.rename(tmp, outPath);
-  } catch (err) {
-    await fs.rm(tmp, { force: true }).catch(() => {});
-    throw err;
-  }
+  await encodeTone({ outPath, freq, metaArgs });
 }
 
 // Summary of what the test suite will see. Derived from FIXTURES so assertions
